rename maia to metrics
