+++ config/navigations/services_navigation.rb
@@ -136,14 +136,14 @@ SimpleNavigation::Configuration.run do |navigation|
     #   storage_nav.dom_attributes = {class: 'content-list'}
     end
 
-    primary.item :resource_management, 'Capacity, Masterdata & Monitoring', nil,
+    primary.item :resource_management, 'Capacity, Masterdata & Metrics', nil,
       html: {class: "fancy-nav-header", 'data-icon': "monitoring-icon" },
       if: -> {services_ng.available?(:resource_management,:resources) or services.available?(:cost_control)} do |monitoring_nav|
       # Disable cost_control and can be deleted after switch to new masterdata api
       #monitoring_nav.item :cost_control,        'Cost Control',        -> {plugin('cost_control').cost_object_path}, if: -> { services.available?(:cost_control) }, highlights_on: Proc.new { params[:controller][/cost_control\/.*/] }
       monitoring_nav.item :audit, 'Audit', -> { plugin('audit').root_path }, if: -> { plugin_available?(:audit) && current_user && current_user.is_allowed?('audit:application_get') }, highlights_on: -> { params[:controller][%r{flavors/?.*}] }
-      monitoring_nav.item :audit, 'Maia', -> { plugin('maia').index_path }, if: -> { plugin_available?(:maia) && current_user && current_user.is_allowed?('maia:show') }, highlights_on: Proc.new { params[:controller][/maia\/.*/] }
       monitoring_nav.item :masterdata_cockpit,  'Masterdata',  -> {plugin('masterdata_cockpit').project_masterdata_path}, if: -> { services_ng.available?(:masterdata_cockpit) }, highlights_on: Proc.new { params[:controller][/masterdata_cockpit\/.*/] }
+      monitoring_nav.item :metrics, 'Metrics', -> { plugin('metrics').index_path }, if: -> { plugin_available?(:metrics) && current_user && current_user.is_allowed?('metrics:show') }, highlights_on: Proc.new { params[:controller][/metrics\/.*/] }
       monitoring_nav.item :resource_management, 'Resource Management ', -> {plugin('resource_management').resources_path}, if: -> { services_ng.available?(:resource_management,:resources) }, highlights_on: Proc.new { params[:controller][/resource_management\/.*/] }
     end
