--- plugins/maia/app/assets/config/maia_manifest.js
@@ -1,2 +1,0 @@
-//= link_directory ../javascripts/maia .js
-//= link_directory ../stylesheets/maia .css
