--- plugins/maia/app/controllers/maia/application_controller.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Maia
-  class ApplicationController < DashboardController
-    def index
-    end
-  end
-end
