--- plugins/maia/app/helpers/maia/application_helper.rb
@@ -1,4 +1,0 @@
-module Maia
-  module ApplicationHelper
-  end
-end
