--- plugins/maia/app/models/maia/application_record.rb
@@ -1,5 +1,0 @@
-module Maia
-  class ApplicationRecord < ActiveRecord::Base
-    self.abstract_class = true
-  end
-end
