--- plugins/maia/app/views/layouts/maia/application.html.erb
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Maia</title>
-  <%= stylesheet_link_tag    "maia/application", media: "all" %>
-  <%= javascript_include_tag "maia/application" %>
-  <%= csrf_meta_tags %>
-</head>
-<body>
-
-<%= yield %>
-
-</body>
-</html>
