--- plugins/maia/config/policy.json
@@ -1,4 +1,0 @@
-{
-  "access_to_project": "not project_id:nil and (role:admin or role:member)",
-  "maia:show": "rule:access_to_project"
-}
