--- plugins/maia/config/routes.rb
@@ -1,3 +1,0 @@
-Maia::Engine.routes.draw do
-  get '/' => 'application#index', as: :index
-end
