--- plugins/maia/lib/maia.rb
@@ -1,5 +1,0 @@
-require "maia/engine"
-
-module Maia
-  # Your code goes here...
-end
