--- plugins/maia/lib/maia/engine.rb
@@ -1,5 +1,0 @@
-module Maia
-  class Engine < ::Rails::Engine
-    isolate_namespace Maia
-  end
-end
