--- plugins/maia/lib/tasks/maia_tasks.rake
@@ -1,4 +1,0 @@
-# desc "Explaining what the task does"
-# task :maia do
-#   # Task goes here
-# end
