--- plugins/maia/test/maia_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class Maia::Test < ActiveSupport::TestCase
-  test "truth" do
-    assert_kind_of Module, Maia
-  end
-end
