+++ plugins/metrics/app/assets/config/metrics_manifest.js
@@ -1,0 +1,2 @@
+//= link_directory ../javascripts/metrics .js
+//= link_directory ../stylesheets/metrics .css
