+++ plugins/metrics/app/assets/stylesheets/metrics/_application.scss
