+++ plugins/metrics/app/controllers/metrics/application_controller.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Metrics
+  class ApplicationController < DashboardController
+    def index
+    end
+  end
+end
