+++ plugins/metrics/app/helpers/metrics/application_helper.rb
@@ -1,0 +1,4 @@
+module Metrics
+  module ApplicationHelper
+  end
+end
