+++ plugins/metrics/app/models/metrics/application_record.rb
@@ -1,0 +1,5 @@
+module Metrics
+  class ApplicationRecord < ActiveRecord::Base
+    self.abstract_class = true
+  end
+end
