+++ plugins/metrics/app/views/layouts/metrics/application.html.erb
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>Metrics</title>
+  <%= stylesheet_link_tag    "metrics/application", media: "all" %>
+  <%= javascript_include_tag "metrics/application" %>
+  <%= csrf_meta_tags %>
+</head>
+<body>
+
+<%= yield %>
+
+</body>
+</html>
