+++ plugins/metrics/app/views/metrics/application/index.html.haml
