+++ plugins/metrics/config/policy.json
@@ -1,0 +1,4 @@
+{
+  "access_to_project": "not project_id:nil and (role:admin or role:member)",
+  "metrics:show": "rule:access_to_project"
+}
