+++ plugins/metrics/config/routes.rb
@@ -1,0 +1,3 @@
+Metrics::Engine.routes.draw do
+  get '/' => 'application#index', as: :index
+end
