+++ plugins/metrics/lib/metrics.rb
@@ -1,0 +1,5 @@
+require "metrics/engine"
+
+module Metrics
+  # Your code goes here...
+end
