+++ plugins/metrics/lib/metrics/engine.rb
@@ -1,0 +1,5 @@
+module Metrics
+  class Engine < ::Rails::Engine
+    isolate_namespace Metrics
+  end
+end
