+++ plugins/metrics/lib/tasks/metrics_tasks.rake
@@ -1,0 +1,4 @@
+# desc "Explaining what the task does"
+# task :metrics do
+#   # Task goes here
+# end
