+++ plugins/metrics/test/integration/navigation_test.rb
