+++ plugins/metrics/test/metrics_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class Metrics::Test < ActiveSupport::TestCase
+  test "truth" do
+    assert_kind_of Module, Metrics
+  end
+end
