+++ plugins/metrics/test/test_helper.rb
